backend/requirements.txt
fastapi
pymongo
python-jose
passlib
pydantic
uvicorn[standard]
bcrypt
python-socketio
asgiref
